Time: 153 ms (31.28%), Space: 70.7 MB (60.91%) - LeetHub

+++ 3066-minimum-operations-to-exceed-threshold-value-ii/3066-minimum-operations-to-exceed-threshold-value-ii.java
@@ -1,12 +1,12 @@
 class Solution {
     public int minOperations(int[] nums, int k) {
-        Queue<Integer> pq = new PriorityQueue<>();
+        Queue<Long> pq = new PriorityQueue<>();
         int count = 0;
         for (int n : nums)
-            pq.offer(n);
+            pq.offer((long) n);
         while (pq.size() >= 2 && pq.peek() < k) {
-            int x = pq.poll(), y = pq.poll();
-            pq.offer(x * 2 + y);
+            long x = pq.poll(), y = pq.poll();
+            pq.offer(x * 2l + y);
             count++;
         }
         return count;
